Vista horario

+++ APITX/app/vistas/vistaHorariodetalle.py
@@ -1,7 +1,9 @@
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-from app.models import TxdHorariodetalle
+from django.core.exceptions import ObjectDoesNotExist
+from datetime import datetime, date
+from app.models import TxdHorariodetalle,TxdBus,TxdChofer
 from app.serializables import TxdHorariodetalleS
 
 
@@ -16,10 +18,31 @@ def lista_objetos(request):
         return Response(serializador.data)
 
     elif request.method == 'POST':
+
         serializador = TxdHorariodetalleS(data=request.data)
         if serializador.is_valid():
+            try:
+                chofer = TxdChofer.objects.get(pk=request.data['chofer'],estado=1)
+            except ObjectDoesNotExist:
+                respuesta ={'crear': {'estado': 'No puede asignar un piloto deshabilitado.'}}
+                return Response(respuesta, status=status.HTTP_406_NOT_ACCEPTABLE)
+
+            try:
+                bus = TxdBus.objects.get(pk=request.data['bus'], estado=1)
+            except ObjectDoesNotExist:
+                respuesta ={'crear': {'estado": "No puede asignar un bus deshabilitado.'}}
+                return Response(respuesta, status=status.HTTP_406_NOT_ACCEPTABLE)
+
+            formato_fecha = "%Y-%m-%d"
+            fecha2 = datetime.strptime(request.data['fecha'], formato_fecha)
+            stringF=str(datetime.now().year)+"-"+str(datetime.now().month)+"-"+str(datetime.now().day)
+            fecha1 = datetime.strptime(stringF, formato_fecha)
+            if fecha2 < fecha1:
+                respuesta ={'crear': {'estado": "La fecha debe ser mayor igual a la fecha actual.'}}
+                return Response(respuesta, status=status.HTTP_406_NOT_ACCEPTABLE)
             serializador.save()
-            return Response(serializador.data, status=status.HTTP_201_CREATED)
+            respuesta ={'crear': {'estado": "Creado Exitosamente'}}
+            return Response(respuesta, status=status.HTTP_201_CREATED)
         return Response(serializador.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET', 'PUT', 'DELETE'])
@@ -29,7 +52,7 @@ def detalle_objetos(request, pk):
     """
     try:
         objeto = TxdHorariodetalle.objects.get(pk=pk)
-    except objeto.DoesNotExist:
+    except ObjectDoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
     if request.method == 'GET':
@@ -39,10 +62,62 @@ def detalle_objetos(request, pk):
     elif request.method == 'PUT':
         serializador = TxdHorariodetalleS(objeto, data=request.data)
         if serializador.is_valid():
+            try:
+                chofer = TxdChofer.objects.get(pk=request.data['chofer'],estado=1)
+            except ObjectDoesNotExist:
+                respuesta ={'modificar': {'estado': 'No puede asignar un piloto deshabilitado.'}}
+                return Response(respuesta, status=status.HTTP_406_NOT_ACCEPTABLE)
+
+            try:
+                bus = TxdBus.objects.get(pk=request.data['bus'], estado=1)
+            except ObjectDoesNotExist:
+                respuesta ={'modificar': {'estado": "No puede asignar un bus deshabilitado.'}}
+                return Response(respuesta, status=status.HTTP_406_NOT_ACCEPTABLE)
+
+            formato_fecha = "%Y-%m-%d"
+            fecha2 = datetime.strptime(request.data['fecha'], formato_fecha)
+            stringF=str(datetime.now().year)+"-"+str(datetime.now().month)+"-"+str(datetime.now().day)
+            fecha1 = datetime.strptime(stringF, formato_fecha)
+            if fecha2 < fecha1:
+                respuesta ={'modificar': {'estado": "La fecha debe ser mayor igual a la fecha actual.'}}
+                return Response(respuesta, status=status.HTTP_406_NOT_ACCEPTABLE)
             serializador.save()
-            return Response(serializador.data)
+            respuesta ={'modificar': {'estado": "Modificado Exitosamente'}}
+            return Response(respuesta, status=status.HTTP_202_ACCEPTED)
         return Response(serializador.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
-        objeto.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        formato_fecha = "%Y-%m-%d"
+
+
+        fecha2 = datetime.combine(objeto.fecha,  datetime.min.time())
+        stringF=str(datetime.now().year)+"-"+str(datetime.now().month)+"-"+str(datetime.now().day)
+        fecha1 = datetime.strptime(stringF, formato_fecha)
+
+        if fecha2 >= fecha1:
+            objeto.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
+        respuesta ={'eliminar': {'estado": "no puede eliminar un registro con una fecha anterior a la actual.'}}
+        return Response(respuesta, status=status.HTTP_406_NOT_ACCEPTABLE)
+
+@api_view(['GET'])
+def rango(request,fInicio,fFin):
+    """
+    Actuliza, elimina un objeto segun su id
+    """
+    formato_fecha = "%Y-%m-%d"
+    inicio = datetime.strptime(fInicio, formato_fecha).date()
+    fin = datetime.strptime(fFin, formato_fecha).date()
+
+    try:
+        hoy = date.today()
+        objeto = TxdHorariodetalle.objects.filter(chofer=1)
+    except ObjectDoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+
+    if request.method == 'GET':
+        serializador = TxdHorariodetalleS(objeto)
+        return Response(serializador.data)
+    else:
+        return Response(status=status.HTTP_404_NOT_FOUND)
